Switch from creating bytecode to a custom proxy
Rather than generating interfaces and abstract classes and letting the default JavaMaker create a
proxy to that, clamp now uses a custom proxy maker that adds its methods directly on the proxy
class.  This cuts out the need for an intermediary proxy class and allows clamp to work properly as
a metatype.